+++ Compose-Flight-Search-App/core/domain/src/main/kotlin/com/example/flightsearchapp/core/domain/DeleteFavoriteUseCase.kt
@@ -1,10 +1,10 @@
-package com.example.flightsearchapp.domain
+package com.example.flightsearchapp.core.domain
 
-import com.example.flightsearchapp.data.FavoritesRepository
-import com.example.flightsearchapp.di.DispatcherDefault
-import javax.inject.Inject
+import com.example.flightsearchapp.core.data.di.DispatcherDefault
+import com.example.flightsearchapp.core.data.repository.FavoritesRepository
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.withContext
+import javax.inject.Inject
 
 class DeleteFavoriteUseCase @Inject constructor(
     private val favoritesRepository: FavoritesRepository,
