+++ Compose-Flight-Search-App/core/domain/src/main/kotlin/com/example/flightsearchapp/core/domain/SetSavedSearchTextUseCase.kt
@@ -1,10 +1,10 @@
-package com.example.flightsearchapp.domain
+package com.example.flightsearchapp.core.domain
 
-import com.example.flightsearchapp.data.SearchTextRepository
-import com.example.flightsearchapp.di.DispatcherDefault
-import javax.inject.Inject
+import com.example.flightsearchapp.core.data.di.DispatcherDefault
+import com.example.flightsearchapp.core.data.repository.SearchTextRepository
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.withContext
+import javax.inject.Inject
 
 class SetSavedSearchTextUseCase @Inject constructor(
     private val searchTextRepository: SearchTextRepository,
